+++ app/models/subscriptions/fee_subscription.rb
@@ -1,0 +1,10 @@
+class FeeSubscription
+  include Subscriptionable
+
+  def set_current_state
+    subscription.state
+  end
+
+  private
+
+end