+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_08_16_134844) do
+ActiveRecord::Schema.define(version: 2020_08_19_173119) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -77,7 +77,9 @@ ActiveRecord::Schema.define(version: 2020_08_16_134844) do
     t.bigint "course_event_id", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
+    t.bigint "subscription_id", null: false
     t.index ["course_event_id"], name: "index_attendees_on_course_event_id"
+    t.index ["subscription_id"], name: "index_attendees_on_subscription_id"
     t.index ["user_id"], name: "index_attendees_on_user_id"
   end
 
@@ -223,8 +225,13 @@ ActiveRecord::Schema.define(version: 2020_08_16_134844) do
     t.integer "days", limit: 2
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
+    t.integer "product_type", limit: 2, null: false
+    t.integer "state", limit: 2
+    t.integer "max_accesses_number", limit: 2
     t.index ["category_id"], name: "index_products_on_category_id"
     t.index ["organization_id"], name: "index_products_on_organization_id"
+    t.index ["product_type"], name: "index_products_on_product_type"
+    t.index ["state"], name: "index_products_on_state"
   end
 
   create_table "rooms", force: :cascade do |t|
@@ -248,11 +255,16 @@ ActiveRecord::Schema.define(version: 2020_08_16_134844) do
     t.date "end_on"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
+    t.integer "subscription_type", limit: 2, null: false
+    t.integer "state", limit: 2
+    t.integer "max_accesses_number", limit: 2
     t.index ["category_id"], name: "index_subscriptions_on_category_id"
     t.index ["code"], name: "index_subscriptions_on_code", unique: true
     t.index ["order_id"], name: "index_subscriptions_on_order_id"
     t.index ["organization_id"], name: "index_subscriptions_on_organization_id"
     t.index ["product_id"], name: "index_subscriptions_on_product_id"
+    t.index ["state"], name: "index_subscriptions_on_state"
+    t.index ["subscription_type"], name: "index_subscriptions_on_subscription_type"
     t.index ["user_id"], name: "index_subscriptions_on_user_id"
   end
 
@@ -311,6 +323,7 @@ ActiveRecord::Schema.define(version: 2020_08_16_134844) do
   add_foreign_key "active_storage_attachments", "active_storage_blobs", column: "blob_id"
   add_foreign_key "admin_users", "organizations"
   add_foreign_key "attendees", "course_events"
+  add_foreign_key "attendees", "subscriptions"
   add_foreign_key "attendees", "users"
   add_foreign_key "categories", "organizations"
   add_foreign_key "course_event_comments", "course_events"