+++ db/seeds.rb
@@ -447,13 +447,13 @@ if CourseEvent.count == 0
 end
 
 if Product.count == 0
-  p_fitness_try   = Product.create!(organization: o_dance, category: ca_fitness, price_cents: 0,     days: 15 , name: 'Periodo di prova'           ,   )
-  p_fitness_month = Product.create!(organization: o_dance, category: ca_fitness, price_cents: 1900,  days: 30 , name: 'Abbonamento mensile fitness',   )
-  p_fitness_year  = Product.create!(organization: o_dance, category: ca_fitness, price_cents: 19900, days: 365, name: 'Abbonamento annuale fitness',   )
-  p_spa_day       = Product.create!(organization: o_dance, category: ca_spa    , price_cents: 2900,  days: 1  , name: 'Entrata singola spa'        ,   )
-  p_spa_month     = Product.create!(organization: o_dance, category: ca_spa    , price_cents: 14900, days: 30 , name: 'Abbonamento mensile spa'    ,   )
-  p_swim_month    = Product.create!(organization: o_swim , category: ca_swim   , price_cents: 4900,  days: 30 , name: 'Abbonamento mensile aquagym',   )
-  p_swim_year     = Product.create!(organization: o_swim , category: ca_swim   , price_cents: 53900, days: 365, name: 'Abbonamento annuale aquagym',   )
+  p_fitness_try   = Product.create!(organization: o_dance, category: ca_fitness, product_type: :trial       , state: :active, price_cents: 0,     days: 15 , name: 'Periodo di prova'           ,   )
+  p_fitness_month = Product.create!(organization: o_dance, category: ca_fitness, product_type: :fee         , state: :active, price_cents: 1900,  days: 30 , name: 'Abbonamento mensile fitness',   )
+  p_fitness_year  = Product.create!(organization: o_dance, category: ca_fitness, product_type: :fee         , state: :active, price_cents: 19900, days: 365, name: 'Abbonamento annuale fitness',   )
+  p_spa_day       = Product.create!(organization: o_dance, category: ca_spa    , product_type: :consumption , state: :active, price_cents: 2900,  days: 1  , name: 'Entrata singola spa'        ,   )
+  p_spa_month     = Product.create!(organization: o_dance, category: ca_spa    , product_type: :consumption , state: :active, price_cents: 14900, days: 30 , name: 'Abbonamento mensile spa'    ,   )
+  p_swim_month    = Product.create!(organization: o_swim , category: ca_swim   , product_type: :fee         , state: :active, price_cents: 4900,  days: 30 , name: 'Abbonamento mensile aquagym',   )
+  p_swim_year     = Product.create!(organization: o_swim , category: ca_swim   , product_type: :fee         , state: :active, price_cents: 53900, days: 365, name: 'Abbonamento annuale aquagym',   )
   puts "Products: #{Product.count}"
 else
   p_fitness_try, p_fitness_month, p_fitness_year, p_spa_day, p_spa_month, p_swim_month, p_swim_year = Product.all
@@ -478,11 +478,11 @@ if Payment.count == 0
   # Redeemed active subscriptions...
  #subscription_attributes << { organization: o_dance, category: p_fitness_year.category , product: p_fitness_year , user: u_stefy, start_on: ord_stefy.start_on, end_on: Time.zone.today + p_fitness_year.days, order: ord_stefy  }
  #subscription_attributes << { organization: o_dance, category: p_spa_month.category    , product: p_spa_month    , user: u_stefy, start_on: ord_stefy.start_on, end_on: Time.zone.today + p_spa_month.days   , order: ord_stefy  }
-  subscription_attributes << { organization: o_dance, category: p_fitness_month.category, product: p_fitness_month, user: u_marco, start_on: Time.zone.today, end_on: Time.zone.today + p_fitness_month.days }
-  subscription_attributes << { organization: o_dance, category: p_fitness_month.category, product: p_fitness_month, user: u_linda, start_on: Time.zone.today, end_on: Time.zone.today + p_fitness_month.days }
-  subscription_attributes << { organization: o_dance, category: p_fitness_try.category  , product: p_fitness_try  , user: u_paolo, start_on: Time.zone.today, end_on: Time.zone.today + p_fitness_try.days   }
+  subscription_attributes << { organization: o_dance, category: p_fitness_month.category, product: p_fitness_month, user: u_marco, subscription_type: :fee  , state: :active, start_on: Time.zone.today }
+  subscription_attributes << { organization: o_dance, category: p_fitness_month.category, product: p_fitness_month, user: u_linda, subscription_type: :fee  , state: :active, start_on: Time.zone.today }
+  subscription_attributes << { organization: o_dance, category: p_fitness_try.category  , product: p_fitness_try  , user: u_paolo, subscription_type: :trial, state: :active, start_on: Time.zone.today }
   generic_users.each do |user|
-    subscription_attributes << { organization: o_dance, category: p_fitness_try.category, product: p_fitness_try  , user: user,    start_on: Time.zone.today, end_on: Time.zone.today + p_fitness_try.days }
+    subscription_attributes << { organization: o_dance, category: p_fitness_try.category, product: p_fitness_try  , user: user,    subscription_type: :fee  , state: :active, start_on: Time.zone.today }
   end
   # Free subscription codes...
   5.times do
@@ -508,8 +508,6 @@ if Attendee.count == 0
   Attendee.new(course_event_id:  3, user_id: u_stefy.id, disable_bookability: true).save!
   Attendee.new(course_event_id:  4, user_id: u_stefy.id, disable_bookability: true).save!
   Attendee.new(course_event_id:  6, user_id: u_stefy.id, disable_bookability: true).save!
-  
-
 
   puts "Attendees: #{Attendee.count}"
 end